+++ project/LibDependencies.scala
@@ -5,7 +5,7 @@ import sbt.ModuleID
 
 object LibDependencies {
   lazy val hmrcFrontendVersion = "1.24.0"
-  private val playFrontendGovukVersion = "0.60.0"
+  private val playFrontendGovukVersion = "0.62.0"
 
   lazy val libDependencies: Seq[ModuleID] = dependencies(
     shared = {